feat(trigger): receive trigger data and prepend to list
jQuery Library has been added to allow for easy DOM manipulation.
`main.js` will hold any JavaScript that will be specific to the app and
includes a `Livestax.on()` function to 'listen' out for any calls to
`newpet`. A list item containing the Pet Name will be then prepended to
the unordered list to provide a history to the user starting with the
most recent at the top.