chore(node.js): convert static to node.js/express.js app
We require a server-side technology and Node.js is an ideal candidate as
it uses a language already used in the tutorials, JavaScript.

On top of Node, we will use Express.js for basic routing, it has been
added using Node Package Manager, and is listed as a dependency in
`package.json`.

To present the files correctly, we will use the ejs templating engine,
also added as a dependency. Ejs was chosen over other such as Jade
because of it's similarity with HTML, requiring little or no changes to
the HTML.

A basic route has been added rendering the index view and the `js`
directory has been moved to a new `public` directory to make it visible
to the application.