feat(signed_request): capture and assign to variable
LiveStax provides a Signed Request as a JSON Web Token (http://jwt.io),
however, to get the token, in Node.js, the parameter body needs to be
parsed with a package called `body-parser`. The package has been added
to `package.json` for easy install using the `npm install` command. The
JSON Web Token will still need to be decoded.

To use the signed request that LiveStax provides, in Node.js this needs
to be parsed with a package called `body-parser`, which has been added
to `package.json`. This token, however, will still need to be decoded.