feat(signed_request): verify and decode
Using the JWT recommended package for verifying a JSON Web Token, the
signed request is verified. However, to verify we require the app
secret that should not be stored in the code, as it is a configurable
variable, therefore we are storing it within a `.env.json` file and it
is imported using the "dot-env" NPM package.