feat(save_history): post pet name and signed request
The Pet name and the signed request are posted to the "addtohistory"
route using jQuery's Post function as it is the library we are
using within this tutorial. This Pet name can then be stored in some
persistent data store against the instance ID once the signed request
token has been decoded.