feat(reset_history): add LiveStax menu item
The pet history is stored against the instance ID of an app so to remove
it we need to delete the data set at the instance ID.

The route to delete the data is called when clicking the "Clear
History" option in the menu. The menu item was added using the
`Livetax.menu.set()`.